Removing dependency on `Value` object.
