+++ main/src/ca/uwaterloo/flix/api/Flix.scala
@@ -26,7 +26,6 @@ import ca.uwaterloo.flix.runtime.quickchecker.QuickChecker
 import ca.uwaterloo.flix.runtime.verifier.Verifier
 import ca.uwaterloo.flix.runtime.{DeltaSolver, Model, Solver, Value}
 import ca.uwaterloo.flix.util.{LocalResource, Options, Validation}
-import org.objectweb.asm.Opcodes.V1_8
 import scala.collection.mutable.ListBuffer
 import scala.collection.{immutable, mutable}
