+++ main/src/ca/uwaterloo/flix/language/phase/CodegenHelper.scala
@@ -426,15 +426,10 @@ object CodegenHelper {
     val flixClass : Class[_] = classOf[Flix]
 
     val unitClass : Class[_] = Value.Unit.getClass
-    val tagClass : Class[_] = classOf[TagInterface]
     val tupleClass : Class[_] = classOf[TupleInterface]
-    val valueObject = "ca/uwaterloo/flix/runtime/Value$"
     val scalaPredef = "scala/Predef$"
     val scalaMathPkg = "scala/math/package$"
     val tagInterface : Class[_] = classOf[TagInterface]
-
-    def loadValueObject(visitor: MethodVisitor): Unit =
-      visitor.visitFieldInsn(GETSTATIC, valueObject, "MODULE$", s"L$valueObject;")
   }
 
   // This constant is used in LoadBytecode, so we can't put it in the private Constants object.
@@ -560,72 +555,4 @@ object CodegenHelper {
     case Expression.MatchError(tpe, loc) => Nil
     case Expression.SwitchError(tpe, loc) => Nil
   }
-
-  /**
-    * Find tuples from the given expression
-    * @param e expression
-    * @return A list containing all the tuples in the given expression
-    */
-  def findTuples(e: Expression): List[Type] = {
-    /**
-      * This method search in the type for a tuple type
-      * @param tpe type to be searched
-      * @return tuple types nested in the type
-      */
-    def searchInType(tpe: Type) : List[Type] = tpe match {
-      case Type.Apply(t, ts) if tpe.isTuple => tpe :: ts.flatMap(searchInType)
-      case _ => Nil
-    }
-
-    e match {
-      case Expression.Unit => Nil
-      case Expression.True => Nil
-      case Expression.False => Nil
-      case Expression.Char(lit) => Nil
-      case Expression.Float32(lit) => Nil
-      case Expression.Float64(lit) => Nil
-      case Expression.Int8(lit) => Nil
-      case Expression.Int16(lit) => Nil
-      case Expression.Int32(lit) => Nil
-      case Expression.Int64(lit) => Nil
-      case Expression.BigInt(lit) => Nil
-      case Expression.Str(lit) => Nil
-      case Expression.LoadBool(n, o) => Nil
-      case Expression.LoadInt8(b, o) => Nil
-      case Expression.LoadInt16(b, o) => Nil
-      case Expression.LoadInt32(b, o) => Nil
-      case Expression.StoreBool(b, o, v) => Nil
-      case Expression.StoreInt8(b, o, v) => Nil
-      case Expression.StoreInt16(b, o, v) => Nil
-      case Expression.StoreInt32(b, o, v) => Nil
-      case Expression.Var(sym, tpe, loc) => searchInType(tpe)
-      case Expression.Ref(name, tpe, loc) => searchInType(tpe)
-      case Expression.MkClosureRef(ref, freeVars, tpe, loc) => searchInType(tpe)
-      case Expression.ApplyRef(name, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
-      case Expression.ApplyTail(name, formals, actuals, tpe, loc) => actuals.flatMap(findTuples) ::: searchInType(tpe)
-      case Expression.ApplyHook(hook, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
-      case Expression.ApplyClosure(exp, args, tpe, loc) => findTuples(exp) ++ args.flatMap(findTuples) ::: searchInType(tpe)
-      case Expression.Unary(op, exp, tpe, loc) => findTuples(exp) ::: searchInType(tpe)
-      case Expression.Binary(op, exp1, exp2, tpe, loc) => findTuples(exp1) ::: findTuples(exp2) ::: searchInType(tpe)
-      case Expression.IfThenElse(exp1, exp2, exp3, tpe, loc) => findTuples(exp1) ::: findTuples(exp2) ::: findTuples(exp3)
-      case Expression.Let(sym, exp1, exp2, tpe, loc) => findTuples(exp1) ::: findTuples(exp2) ::: searchInType(tpe)
-      case Expression.Is(sym, tag, exp, loc) => findTuples(exp)
-      case Expression.Tag(enum, tag, exp, tpe, loc) => findTuples(exp) ::: searchInType(tpe)
-      case Expression.Untag(sym, tag, exp, tpe, loc) => findTuples(exp) ::: searchInType(tpe)
-      case Expression.Index(base, offset, tpe, loc) => findTuples(base) ::: searchInType(tpe)
-      case Expression.Tuple(elms, tpe, loc) => elms.flatMap(findTuples).toList ++ List(tpe) ::: searchInType(tpe)
-      case Expression.Reference(exp, tpe, loc) => ??? // TODO
-      case Expression.Dereference(exp, tpe, loc) => ??? // TODO
-      case Expression.Assignment(exp1, exp2, tpe, loc) => ??? // TODO
-      case Expression.Existential(params, exp, loc) => findTuples(exp)
-      case Expression.Universal(params, exp, loc) => findTuples(exp)
-      case Expression.NativeConstructor(constructor, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
-      case Expression.NativeField(field, tpe, loc) => searchInType(tpe)
-      case Expression.NativeMethod(method, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
-      case Expression.UserError(tpe, loc) => searchInType(tpe)
-      case Expression.MatchError(tpe, loc) => searchInType(tpe)
-      case Expression.SwitchError(tpe, loc) => searchInType(tpe)
-    }
-  }
-
 }
