+++ main/src/ca/uwaterloo/flix/language/phase/TupleGen.scala
@@ -18,6 +18,7 @@ package ca.uwaterloo.flix.language.phase
 
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.CompilationError
+import ca.uwaterloo.flix.language.ast.ExecutableAst.Expression
 import ca.uwaterloo.flix.language.ast.{ExecutableAst, Type}
 import ca.uwaterloo.flix.util.{Evaluation, InternalCompilerException, Validation}
 import ca.uwaterloo.flix.util.Validation._
@@ -25,10 +26,11 @@ import ca.uwaterloo.flix.language.phase.CodegenHelper._
 import org.objectweb.asm
 import org.objectweb.asm.{ClassWriter, Label}
 import org.objectweb.asm.Opcodes._
+
 object TupleGen extends Phase[ExecutableAst.Root, ExecutableAst.Root]{
 
   /**
-    * Generate bytecode for Tuples.
+    * Generate bytecode for tuples.
     * The steps that we take to generate enums is as follows:
     *
     * 1. Extract all tuple types from definitions.
@@ -460,5 +462,72 @@ object TupleGen extends Phase[ExecutableAst.Root, ExecutableAst.Root]{
     // Parameters of visit max are thrown away because visitor will calculate the frame and variable stack size
     method.visitMaxs(1, 1)
     method.visitEnd()
+  }
+
+  /**
+    * Find tuples from the given expression
+    * @param e expression
+    * @return A list containing all the tuples in the given expression
+    */
+  def findTuples(e: Expression): List[Type] = {
+    /**
+      * This method search in the type for a tuple type
+      * @param tpe type to be searched
+      * @return tuple types nested in the type
+      */
+    def searchInType(tpe: Type) : List[Type] = tpe match {
+      case Type.Apply(t, ts) if tpe.isTuple => tpe :: ts.flatMap(searchInType)
+      case _ => Nil
+    }
+
+    e match {
+      case Expression.Unit => Nil
+      case Expression.True => Nil
+      case Expression.False => Nil
+      case Expression.Char(lit) => Nil
+      case Expression.Float32(lit) => Nil
+      case Expression.Float64(lit) => Nil
+      case Expression.Int8(lit) => Nil
+      case Expression.Int16(lit) => Nil
+      case Expression.Int32(lit) => Nil
+      case Expression.Int64(lit) => Nil
+      case Expression.BigInt(lit) => Nil
+      case Expression.Str(lit) => Nil
+      case Expression.LoadBool(n, o) => Nil
+      case Expression.LoadInt8(b, o) => Nil
+      case Expression.LoadInt16(b, o) => Nil
+      case Expression.LoadInt32(b, o) => Nil
+      case Expression.StoreBool(b, o, v) => Nil
+      case Expression.StoreInt8(b, o, v) => Nil
+      case Expression.StoreInt16(b, o, v) => Nil
+      case Expression.StoreInt32(b, o, v) => Nil
+      case Expression.Var(sym, tpe, loc) => searchInType(tpe)
+      case Expression.Ref(name, tpe, loc) => searchInType(tpe)
+      case Expression.MkClosureRef(ref, freeVars, tpe, loc) => searchInType(tpe)
+      case Expression.ApplyRef(name, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
+      case Expression.ApplyTail(name, formals, actuals, tpe, loc) => actuals.flatMap(findTuples) ::: searchInType(tpe)
+      case Expression.ApplyHook(hook, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
+      case Expression.ApplyClosure(exp, args, tpe, loc) => findTuples(exp) ++ args.flatMap(findTuples) ::: searchInType(tpe)
+      case Expression.Unary(op, exp, tpe, loc) => findTuples(exp) ::: searchInType(tpe)
+      case Expression.Binary(op, exp1, exp2, tpe, loc) => findTuples(exp1) ::: findTuples(exp2) ::: searchInType(tpe)
+      case Expression.IfThenElse(exp1, exp2, exp3, tpe, loc) => findTuples(exp1) ::: findTuples(exp2) ::: findTuples(exp3)
+      case Expression.Let(sym, exp1, exp2, tpe, loc) => findTuples(exp1) ::: findTuples(exp2) ::: searchInType(tpe)
+      case Expression.Is(sym, tag, exp, loc) => findTuples(exp)
+      case Expression.Tag(enum, tag, exp, tpe, loc) => findTuples(exp) ::: searchInType(tpe)
+      case Expression.Untag(sym, tag, exp, tpe, loc) => findTuples(exp) ::: searchInType(tpe)
+      case Expression.Index(base, offset, tpe, loc) => findTuples(base) ::: searchInType(tpe)
+      case Expression.Tuple(elms, tpe, loc) => elms.flatMap(findTuples).toList ++ List(tpe) ::: searchInType(tpe)
+      case Expression.Reference(exp, tpe, loc) => ??? // TODO
+      case Expression.Dereference(exp, tpe, loc) => ??? // TODO
+      case Expression.Assignment(exp1, exp2, tpe, loc) => ??? // TODO
+      case Expression.Existential(params, exp, loc) => findTuples(exp)
+      case Expression.Universal(params, exp, loc) => findTuples(exp)
+      case Expression.NativeConstructor(constructor, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
+      case Expression.NativeField(field, tpe, loc) => searchInType(tpe)
+      case Expression.NativeMethod(method, args, tpe, loc) => args.flatMap(findTuples) ::: searchInType(tpe)
+      case Expression.UserError(tpe, loc) => searchInType(tpe)
+      case Expression.MatchError(tpe, loc) => searchInType(tpe)
+      case Expression.SwitchError(tpe, loc) => searchInType(tpe)
+    }
   }
 }
